# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HDMI Multiviewer Proxy",
    "description": "REST API for controlling UHD-401MV multiviewer via RS-232",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/multiview": {
      "get": {
        "tags": [
          "Display"
        ],
        "summary": "Get Multiview",
        "description": "Get current multiview display mode.",
        "operationId": "get_multiview_api_multiview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiviewResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Display"
        ],
        "summary": "Set Multiview",
        "description": "Set multiview display mode.",
        "operationId": "set_multiview_api_multiview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultiviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/windows/{window_id}": {
      "get": {
        "tags": [
          "Display"
        ],
        "summary": "Get Window Input",
        "description": "Get input source for a specific window.",
        "operationId": "get_window_input_api_windows__window_id__get",
        "parameters": [
          {
            "name": "window_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 4,
              "minimum": 1,
              "description": "Window ID (1-4)",
              "title": "Window Id"
            },
            "description": "Window ID (1-4)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WindowInput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/windows/{window_id}/input": {
      "post": {
        "tags": [
          "Display"
        ],
        "summary": "Set Window Input",
        "description": "Set input source for a specific window.",
        "operationId": "set_window_input_api_windows__window_id__input_post",
        "parameters": [
          {
            "name": "window_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 4,
              "minimum": 1,
              "description": "Window ID (1-4)",
              "title": "Window Id"
            },
            "description": "Window ID (1-4)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetWindowInputRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WindowInput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/input": {
      "get": {
        "tags": [
          "Display"
        ],
        "summary": "Get Input Source",
        "description": "Get current input source (single screen mode).",
        "operationId": "get_input_source_api_input_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputSourceResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Display"
        ],
        "summary": "Set Input Source",
        "description": "Set input source (single screen mode).",
        "operationId": "set_input_source_api_input_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetInputSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputSourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pip": {
      "get": {
        "tags": [
          "Display"
        ],
        "summary": "Get Pip Settings",
        "description": "Get PIP window settings.",
        "operationId": "get_pip_settings_api_pip_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PIPResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Display"
        ],
        "summary": "Set Pip Settings",
        "description": "Set PIP window settings.",
        "operationId": "set_pip_settings_api_pip_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPIPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PIPResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pbp": {
      "get": {
        "tags": [
          "Display"
        ],
        "summary": "Get Pbp Settings",
        "description": "Get PBP window settings.",
        "operationId": "get_pbp_settings_api_pbp_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PBPResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Display"
        ],
        "summary": "Set Pbp Settings",
        "description": "Set PBP window settings.",
        "operationId": "set_pbp_settings_api_pbp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPBPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PBPResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/triple": {
      "get": {
        "tags": [
          "Display"
        ],
        "summary": "Get Triple Settings",
        "description": "Get triple screen settings.",
        "operationId": "get_triple_settings_api_triple_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripleQuadResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Display"
        ],
        "summary": "Set Triple Settings",
        "description": "Set triple screen settings.",
        "operationId": "set_triple_settings_api_triple_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTripleQuadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripleQuadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/quad": {
      "get": {
        "tags": [
          "Display"
        ],
        "summary": "Get Quad Settings",
        "description": "Get quad screen settings.",
        "operationId": "get_quad_settings_api_quad_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripleQuadResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Display"
        ],
        "summary": "Set Quad Settings",
        "description": "Set quad screen settings.",
        "operationId": "set_quad_settings_api_quad_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTripleQuadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripleQuadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InputSourceResponse": {
        "properties": {
          "input": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input"
          },
          "input_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Name"
          }
        },
        "type": "object",
        "required": [
          "input"
        ],
        "title": "InputSourceResponse",
        "description": "Single screen input source response."
      },
      "MultiviewModeEnum": {
        "type": "string",
        "enum": [
          "single",
          "pip",
          "pbp",
          "triple",
          "quad"
        ],
        "title": "MultiviewModeEnum",
        "description": "Multiview display modes."
      },
      "MultiviewRequest": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/MultiviewModeEnum"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "MultiviewRequest",
        "description": "Multiview mode request."
      },
      "MultiviewResponse": {
        "properties": {
          "mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MultiviewModeEnum"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "MultiviewResponse",
        "description": "Multiview mode response."
      },
...
      "PBPResponse": {
        "properties": {
          "mode": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "aspect": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aspect"
          }
        },
        "type": "object",
        "title": "PBPResponse",
        "description": "PBP settings response."
      },
      "PIPResponse": {
        "properties": {
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "size": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          }
        },
        "type": "object",
        "title": "PIPResponse",
        "description": "PIP settings response."
      },
...
      "SetInputSourceRequest": {
        "properties": {
          "input": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Input"
          }
        },
        "type": "object",
        "required": [
          "input"
        ],
        "title": "SetInputSourceRequest",
        "description": "Set single screen input request."
      },
...
      "SetPBPRequest": {
        "properties": {
          "mode": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "aspect": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Aspect",
            "description": "1=full_screen, 2=16:9"
          }
        },
        "type": "object",
        "title": "SetPBPRequest",
        "description": "Set PBP settings request."
      },
      "SetPIPRequest": {
        "properties": {
          "position": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 4.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Position",
            "description": "1=left_top, 2=left_bottom, 3=right_top, 4=right_bottom"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Size",
            "description": "1=small, 2=middle, 3=large"
          }
        },
        "type": "object",
        "title": "SetPIPRequest",
        "description": "Set PIP settings request."
      },
...
      "SetTripleQuadRequest": {
        "properties": {
          "mode": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "aspect": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Aspect",
            "description": "1=full_screen, 2=16:9"
          }
        },
        "type": "object",
        "title": "SetTripleQuadRequest",
        "description": "Set Triple/Quad settings request."
      },
...
      "SetWindowInputRequest": {
        "properties": {
          "input": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Input"
          }
        },
        "type": "object",
        "required": [
          "input"
        ],
        "title": "SetWindowInputRequest",
        "description": "Set window input request."
      },
      "TripleQuadResponse": {
        "properties": {
          "mode": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "aspect": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aspect"
          }
        },
        "type": "object",
        "title": "TripleQuadResponse",
        "description": "Triple/Quad settings response."
      },
...
      "WindowInput": {
        "properties": {
          "window": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Window"
          },
          "input": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Input"
          },
          "input_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Name"
          }
        },
        "type": "object",
        "required": [
          "window",
          "input"
        ],
        "title": "WindowInput",
        "description": "Window input mapping."
      },
...
    }
  }
}